refactor EdgeConvertFileParser, separate EdgeFileParser to own class
